feat: add a settings modal to the web client (#140)
Added a new modal implementation along with a new settings modal.

## Implementation Details
### Key Changes
- A modal management system was included into the web client. Adding new
modal types is super easy with this system.
- The settings modal was added using the mentioned modal management
system.
- Minor refactoring was done for easier readability in the component
tree.
- The journal service had a compilation error and was fixed here (I need
to hurry up with #135 to prevent this...).

### Design Choices
- Instead of using parallel routes as before, modals are implemented
using a global state that controls the visibility of a component in the
root layout of the web client.
- I did this because I thought this would enable a much easier
experience for central modal management using simple hooks. I don't see
any problems with this now, but I'm not entirely sure about the
downsides of this approach at this point. The components work fine and
they are already split into separate files, so going with another
approach won't be that hard.