feat: add centralized icon management system in web client (#144)
Added a centralized icon management system to the web client.
Previously, the web client pulled icons from various sources:
lucide-icons and react-icons. This has a lot of problems since I can't
directly manage icons and consistency is hard to manage as well. In this
pull request, I used a sprite sheet of icons and a single Icon component
for easier management.

## Implementation Details
### Key Changes
- Added an Icon component that pulls from a SVG sprite sheet for easier
management.
- Added storybook for better component management (I only added stories
for the Icon component for now).

### Dependencies
- The storybook dependency was newly added.
- Lucide-icons and React-icons were removed from the dependencies.

## Additional Context
### Related Issues
- Closes #141